+++ ExaminationSystem_API/ExaminationSystem_API/Dto/DepartmentDTO/DepartmentLookupDTO.cs
@@ -2,7 +2,7 @@ namespace ExaminationSystem_API.Dto.DepartmentDTO
 {
     public class DepartmentLookupDTO
     {
-        public int DepartmentId { get; set; }
+        public byte DepartmentId { get; set; }
         public string DepartmentName { get; set; }
     }
 }
