+++ ExaminationSystem_API/ExaminationSystem_API/Dto/DepartmentDTO/UpdateDepartmentDTO.cs
@@ -5,6 +5,6 @@ namespace ExaminationSystem_API.Dto.DepartmentDTO
     public class UpdateDepartmentDTO :BaseDepartmentDTO
     {
         [Required(ErrorMessage = "Department ID is required for update")]
-        public int DeptId { get; set; }
+        public byte DeptId { get; set; }
     }
 }
