+++ ExaminationSystem_API/ExaminationSystem_API/Repository/ClassRepository/DepartmentRepository.cs
@@ -9,9 +9,9 @@ namespace ExaminationSystem_API.Repository.ClassRepository
         {
             _context = context;
         }
-        public async Task AddDepartmentWithStoredAsync(string name, int BranchID) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_AddDepartment @Deptname={name},@BranchId ={BranchID} ");
-        public async Task UpdateDepartmentWithStoredAsync(int DeptId , string name, int BranchID) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_UpdateDepartment @DeptId ={DeptId},@DeptName={name},@BranchId ={BranchID}");
-        public async Task DeleteDepartmentWithStoredAsync(int DeptId )  => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_DeleteDepartment @DeptId={DeptId}");
+        public async Task AddDepartmentWithStoredAsync(string name, byte BranchID) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_AddDepartment @Deptname={name},@BranchId ={BranchID} ");
+        public async Task UpdateDepartmentWithStoredAsync(byte DeptId , string name, byte BranchID) => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_UpdateDepartment @DeptId ={DeptId},@DeptName={name},@BranchId ={BranchID}");
+        public async Task DeleteDepartmentWithStoredAsync(byte DeptId )  => await _context.Database.ExecuteSqlInterpolatedAsync($"EXEC [TrainingMangerStp].stp_DeleteDepartment @DeptId={DeptId}");
 
     }
 }
