+++ ExaminationSystem_API/ExaminationSystem_API/Repository/InterfaceRepository/ITrackRepository.cs
@@ -2,7 +2,7 @@ namespace ExaminationSystem_API.Repository.InterfaceRepository
 {
     public interface ITrackRepository :IGenericRepository<Track>
     {
-        Task AddTrackWithStoredAsync(string name, int deptId);
+        Task AddTrackWithStoredAsync(string name, byte deptId);
         Task UpdateTrackWithStoredAsync(short trackId, string name, byte deptId);
     }
 }
